requirements.txt
nicegui>=1.4
clickhouse-driver>=0.2.6
python-dotenv>=1.0
pytest>=7.0
